requirements.txt
accelerate==1.3.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.2.1
diffusers==0.32.2
dnspython==2.7.0
email_validator==2.2.0
filelock==3.17.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.2
fsspec==2025.2.0
greenlet==3.1.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
moviepy==2.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==10.4.0
proglog==0.1.10
psutil==6.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
razorpay==1.4.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
SQLAlchemy==2.0.37
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
